TakeNews/requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.4
django-celery-beat==2.5.0
django-filter==23.2
django-timezone-field==5.1
djangorestframework==3.14.0
docker==6.1.3
exceptiongroup==1.1.3
flower==2.0.1
h11==0.14.0
humanize==4.8.0
idna==3.4
kombu==5.3.1
Markdown==3.4.4
outcome==1.2.0
packaging==23.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tornado==6.3.3
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.1
wsproto==1.2.0